requirements.txt
pandas==2.2.2
pyarrow==17.0.0   
requests==2.32.3
pyyaml==6.0.2
pytest==8.3.3
